MAJOR CHANGE: remapping <leader> from ',' to ' '
The reasoning behind this is to be able to move back fast when
overstepping a search within a line using fFtT.

Previous leader (','), made the program wait some milliseconds to allow
for the intepretation of a command that could follow.

Now that no command can follow ',', the backtracking is faster.